feat: add useBatchAnalysis hook for batch analysis of shogi games
- Implemented useBatchAnalysis hook to manage batch analysis state and operations.
- Added functions for analyzing specific plies, nodes, and branches.
- Integrated engine pool for parallel analysis jobs.
- Enhanced error handling and messaging for analysis processes.

refactor: expose EnginePoolHandle and UseKifuNavigationResult interfaces

- Changed EnginePoolHandle and UseKifuNavigationResult interfaces to be exported for better accessibility.

feat: create useMoveExecution hook for handling move execution

- Developed useMoveExecution hook to manage move execution logic in various modes (edit, review, game).
- Implemented handlers for square selection, promotion choices, and hand selections.
- Integrated legal move fetching and validation logic.
